moved user stories and checklist for project part 4 to separated.md file as per TA feedback, so we can remove them from readme.md
moved user stories and risk assessments for project part 4, along with a checklist for tracking completion to a new file, so we can remove them from readme.md as per TA's feedback. Right now we have it as .md file in the root, however, once we create a separate directory for part4 after the completion of part3, I will move it to that directory.